+++ infra/examples-dev/aws-google-workspace/vpc-variables.tf
@@ -1,5 +1,0 @@
-variable "vpc_ip_block" {
-  type        = string
-  description = "IP block for VPC to create for psoxy instances, in CIDR notation"
-  default     = "10.0.0.0/18"
-}