+++ CHANGELOG.md
@@ -15,6 +15,8 @@ BREAKING:
   - `azuread-local-cert` module variables have changed; you must now pass `application_id` instead
     of `application_object_id`; these refer to different values you can obtain via the Entra ID web
     portal (formally Azure AD portal)
+  - variables to `aws-host`/`gcp-host` modules to have changed slightly; if you initially copied an
+    example based on 0.4.x, you may have to update some variable names in your `main.tf`.
 
 ## [0.4.61](https://github.com/Worklytics/psoxy/release/tag/v0.4.61)
  - added some `columnsToPseudonymizeIfPresent` to survey bulk connectors; these are to avoid PII